Cleanup component organization
