+++ packages/fluent/src/components/index.ts
@@ -3,6 +3,10 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 export * from './ColorPicker.js'
+export * from './ColorPicker.types.js'
 export * from './ColorPickerButton.js'
-export * from './ScaleDropdown/index.js'
+export * from './ColorPickerButton.types.js'
+export * from './ScaleDropdown/ScaleDropdown.js'
+export * from './ScaleDropdown/ScaleDropdown.types.js'
 export * from './ScaleTypeChoiceGroup.js'
+export * from './ScaleTypeChoiceGroup.types.js'
